The student can show his grade after complete his exam, and he can not access that exam again

+++ College examination management system/Software2 project/Models/ExamModel.cs
@@ -14,6 +14,6 @@ namespace Software2_project.Models
         [Required]
         public short course_id { get; set; }
         [Required]
-        public short grade { get; set; }
+        public float grade { get; set; }
     }
 }
